Add user registration middleware
added register() method in ./controllers/user to handle inserting the new user if user is not found by email.

bcrypt is used to salt and hash the plain text password that is passed through the request.

In the future, the security of the registration system can be locked down further by introducing some middleware to check against multiple registration requests coming from the same ip.

Task# 51