Add login middleware: controller/auth
The login features needs to check for an existing user, validate the password and store a user session somehwere on the client.

bcrypt.compare() is used to check the password.

jsonwebtoken used to generate a token containing a payload of the user's data given some secret key.

Additional auth middleware functions should be added to better secure the login system.

ie:
 - limiting login attempts
 - checking ip on login attempts
 - 2 factor authetnication

Task# 52