Added ./client/ to store react.js app
reactjs is used for rendering/managing the views for the application.

React isn't specific to express.js, so it doesn't necessarily need to reside on the same server and/or project.
However, for this application, the reactjs app will be served by the same express.js server that handles all of our backend functionality.

Ideally, the reactjs part of the application should be separated from our backend application to give greater flexability when scaling and managing the chat application in the future.

For now, the two will be maintained in the same repo for convenience.

Task# 53