Add proxy server to client/package.json
The proxy option in package.json allows us to set a reference to the express server, enabling us to make ajax calls using relative paths vs absolute paths.

create-react-app defaults to running a dev server on port: 3000, however, our express server is set to run on port 7777.

If the client and backend resided on two separate servers, the proxy option would not be necessary in such a case.

Task# 53