Add redux for client side state management
Client side applications typically go through a series of interactions between the application and the user.

ie:
 - login
 - update profile
 - search query
 - comment on a blog post

Redux provides a means to track these changes within the application.
The library's emphasis on immutability allows for the ability to "time travel" and step through changes.

Redux Terminology:

 Store

 - Refers to the data set being used/tracked inside of the application

   ie:
     - users
     - chat messages
     - logged in/out

 Actions

 - Events that occur throughout the application that may alter the state of the application.
 - Basically, actions decide which reducers the data needs to pass through to update the application state.

 ie:
  - USER_LOGGED_IN

  - USER_LOGGED_OUT

 Reducers

 - The duty of a reducer is to update the data in the application state.
 - reducer's accept some data commonly referred as action.payload and returns the updated version of state.

Redux Thunk middleware enables the ability return functions in redux actions vs plain object actions.
Essentially, it allows for more granular controls on when a redux action is run.

ie: You need to fetch some api data, and the action to be dispatched is dependent on the result of the api call.