Add loginUser action
Added ./actions/auth.js to store user authentication related redux actions

If loginUser successfully communicates with the backend server and the user data is verified, the
backend server will respond with a jsonwebtoken, which will be stored client side for a user session.
Dispatching a USER_LOGGED_IN action will be added later to update the redux store w/ the currently logged in user's data.

Task# 56