Maintain user session by setting jsonwebtoken on successful login
jsonwebtokens gives us a way to persist a user's session after a successful login.

The token consists of some payload/data about the currently logged in user.

In our case, the jsonwebtoken is being stored in the browser's localStorage. However, there are other methods that may be explored inthe future such as storing the token in a browser cookie and/or server side.

TODO: Clear the token on logout.

Task# 57