Add PrivateRoute component and set ChatRoom as private route.
A PrivateRoute component is a method that enables privatizing certain Routes/Components based onsome condition(s). In our case, PrivateRoute checks if the client trying to access the route is authenticated - they have logged in. The client is either redirected back to the login page and/or to the requested route.

ChatRoom, the main Component for the chat application was set as a PrivateRoute, because it's noot meant to be a public facing app.

Task# 59