Add message submit functionality
ChatWindow component holds the input field for message submissions.

Controlled message input
 - the input field's value is set to the component's state.input
 - the state.input is updated on the the onChage

Message Submit
onKeyPress is used to to "submit" the message on "Enter" key press and subsequentially clearing state.input to empty the input field.
At this point, the message input submission is visually what we expect, but the backend and integration with socket.io will be developed later.

Task# 62