Update a user's socketId field in mongodb after a user logs in
a client/user's current SocketId is updated in mongodb when a user logs in, so that the server's socket has a reference to communicate with the client directly.

When a socket disconnects, we need to broadcast to the rest of the clients/sockets that the user left. Our only reference on socket.io's disconnet event is the socket. We can use the socket.id on disconnect to find the user.

In the future, we can move socketId to be stored in memory using something like redis instead.

Task# 62