Add initChatService to fetch chat data to serve to the client
When a user logs in we need to fetch the chat data in order to start the chat.

A user is redirected to the ChatRoom component after login.
When the commponent mounts, a socket.emit(CHAT_SERVICE_START)
event is sent to request the chat app data from the server.

On CHAT_SERVICE_START, the server socket fetches the data and responds with the data by running the callback function sent by the client.

Data:

users - An object containing all of the users in the application.

rooms - All of the rooms the current user is in.

room - The active room.

messages - All of the messages in the activeroom.

TODO: update the client store with the chat data.

Task# 65