Add newly created room to the client's store
When a room is created, the client that initiated the creation needs to update to reflect the newly added room.
Subsequently, the user's in the room also need to update their client's store.

TODO:
auto open the room for the client that created the room.
add feature to join/change rooms by clicking the room in the sidebar.

Task# 68