Fixed direct message search room open/create
Direct Message Search functionality was broken b/c the logic that determines whether or not to create a room or open a room
was missing a userId parameter.

ie:

getDirectMessage() needed the userid param and needed to check if the currently clicked user is in any of the direct messages.

Task# 74