Add UserMeta Model
Prior to adding UserMeta model, userMetaData was being stored directly in the User model.
Storing a metaData field in the User model made it difficult to manage updating the data in the db
as well as in the redux store.

UserMeta model was created to store additional data specific to the user.

ie:

 - room data
   - unread messages
   - customizations etc.

 - preferences
 - settings

Added UserMeta sample data.

Task# 73