Add addUnreadMessage socket action
Used in parallel with the redux addUnreadMessage action.
It's triggered when a message is submitted and the current client
is not actively in the room of the message.

Emits a MESSAGE_ADD_UNREAD event to the server side socket in order
to update the UserMeta to incremement the unread message count for room.

Task# 73