Update on MESSAGE_ADDED to run addUnreadMessage if user is not actively in the room
The default behavior for on MESSAGE_ADDED event is to simply add the message to the room,
so the ChatWindow updates with the message that was broadcasted to all of the clients.

However, if the user is not actively in the room, we want to mark the message in the room as unread
by firing off addUnreadMessage redux and socket actions.

Task# 73