Add resetMessageCount and IncrementMessageCount to currentUser reducer
When a message is unread or MESSAGE_ADD_UNREAD is triggered, we want to
increment the UserMeta unreadMessageCount by 1 for the message's room.

When a user opens a room or ROOM_OPEN, we want to reset the unreadMEssageCount
back to 0 b/c the user "read" the messages if they opened the room.

Task# 73