Add registerUser action and run on register form submit
regiserUser() posts to the the /api/v1/register endpoint on the express server.

The endpoint's middleware handles registration validation and inserting the user into the db.

Server responds with status 500 on failure w/ errors and 200 and succes on success.

TODO:

Add dispatch redux action to handle notifying user of the registration error(s).

Fix redirect to /login on successful registration.

Task# 75