Add errors reducer and display error messages registration errors
An errors reducer was added to hold general application errors.

The redux store error key is updated with the registration forms errors when there is a failed registerUser()
redux action.

The RegisterForm component displays the errors on re-render if an error is detected in the store.

TODO: clear the errors in the RegisterForm state on input change.

Task# 75