Add rendering of LoginForm errors
Feature is nearly identical to rendering of errors in Registerform.

Hold the errors in component state.

Update the state.errors if login error is found in store.errors which renders the error message on re-render.

Clear the errors on input change.

In the future, a system should be developed to better handle/render error messages in the application.

ie:
Functions can be written to enforce error message structures from the server which will allow the front end to better handle
rendering if the error messages.

Task# 76