Fix broken re-render of direct messages after added direct message
BUG DESCRIPTION:

When a direct message is added, all of the users in the direct mesage list are replaced
by the current user's name.

The current user's name at the top of the sidebar is also empty on re-render.

SOLUTION:

A currentUser reducer was added includes an action ROOM_OPEN that is triggered when a current user
selects a user from the direct message search list.

The role of updating the currentUser on ROOM_OPEN is to clear/reset the unreadMessageCount for that room.
However, the redux actions that were triggering ROOM_OPEN for direct messages were not adding the currentUser
into the payload.

Therefore, an undefined currentUser in the currrentUser reducer was the culprit.

Task# 80