Send NODE_ENV to the client and create socket io connection based on env
The development environment for this project is slightly different from the production.

The development env creates 2 test servers.

 - 1 for the client ( react )
 - 1 for the server side logic (express, node)

The production env runs 1 server

 - the server serves the react application in production.

TEMP FIX:

Send the NODE_ENV from the server to the client on user login, and when the client attempts to create the socket connection,
if in development, we want to explicitly declare the express/node server's url.

If we don't, io() will try to create the socket connection using the react app server which does not work b/c all of the server side socket logic resides elsewhere.

Task# 98