Fix disconnected user status update
On the server socket disconnect event, a USER_DISCONNECTED event is broadcasted to the rest of the clients.

Added redux action setUserDisconnected() to update the user in the store.

Client socket on USER_DISCONNECTED event - added to ChatRoom and runs setUserDisconnect() to update the logged out user's
status on the client.

Task# 91