Add getRoomMetaKey as a static fn on UserMeta model
Updated usage in UserMeta.

The method was added to the UserMeta model for use on the server side.
The previous commit added the function for use on the client side.

In the future, generic functions like this should be exported into an npm module added as a dependency to the client/server instead.

Task# 101